feat(omada): make the crawler self-contained via the wizard plugin system
Moves the Omada IGA crawler's UI wizard, type-picker entry, and live
$metadata discovery out of shared CrawlersPage.jsx/jobs.js and into
tools/crawlers/omada/ (CrawlerMeta.js, ConfigWizard.jsx, discover.js),
mirroring the midPoint crawler's migration. No functional change — same
wizard steps, fields, and validation behavior.

- tools/crawlers/omada/CrawlerMeta.js — type-picker registration
- tools/crawlers/omada/ConfigWizard.jsx — extracted from CrawlersPage.jsx's
  inline OmadaWizard, now using the shared Stepper component for consistency
  with other crawler wizards
- tools/crawlers/omada/discover.js — ported from the removed
  POST /admin/omada/validate-metadata route, now using db.queryOne instead
  of the legacy mssql-shim pool pattern
- Removed app/api/src/omada/metadataProxy.js (folded into discover.js,
  matching midPoint's pattern); updated the CodeQL paths-ignore entry
  accordingly
- Removed 'omada' from the PENDING_MIGRATION lists in pr.yml's
  crawler-manifest guardrail and CrawlerManifest.Tests.ps1
- Ported the validate-metadata test coverage to a new
  app/api/src/routes/omadaDiscover.test.js that exercises discover.js
  directly
