+++ .github/codeql/codeql-config.yml
@@ -6,13 +6,13 @@ paths-ignore:
   # cannot silence without the GitHub Advanced Security inline-suppression setting.
   - app/api/src/llm/scraper.js
 
-  # metadataProxy.js fetches the Omada OData $metadata endpoint on behalf of the
-  # admin wizard. The URL is admin-supplied (Omada can be on-prem or SaaS) and
-  # scheme-validated to http/https before reaching this file. The calling endpoint
-  # is protected by admin auth middleware.
-  - app/api/src/omada/metadataProxy.js
+  # tools/crawlers/omada/discover.js fetches the Omada OData $metadata endpoint
+  # on behalf of the admin wizard. The URL is admin-supplied (Omada can be
+  # on-prem or SaaS) and scheme-validated to http/https before reaching this
+  # file. The calling endpoint is protected by admin auth middleware.
+  - tools/crawlers/omada/discover.js
 
   # tools/crawlers/midpoint/discover.js fetches the midPoint REST API (token
   # endpoint + object search) on behalf of the admin wizard. Same rationale as
-  # metadataProxy.js: admin-supplied URL, scheme-validated, admin auth required.
+  # omada/discover.js: admin-supplied URL, scheme-validated, admin auth required.
   - tools/crawlers/midpoint/discover.js
