+++ tools/crawlers/omada/CrawlerMeta.js
@@ -1,0 +1,5 @@
+export default {
+  id: 'omada',
+  name: 'Omada IGA',
+  description: 'Sync business roles, identities, role assignments, and certification reviews directly from the Omada REST API',
+};
